Create patch set 2
Uploaded patch set 2: Commit message was updated.

Patch-set: 2
Subject: Reverting install_k8s and install_contrail because we want to add support for Kubernetes as the schedule of Contrail containers, which means that k8s needs to be installed before. It is basically the same logic as in Openstack.
Commit: 135e1ea90fe351009ac25d297c71bb50f870f7bc