Update p-t-f-a-d role with any nodeset support.
This change will allow the p-t-f-a-d role to support
any nodeset for kolla/k8s setups. All the nodes in the nodeset
for the job need to be assigned to ansible groups in the job
configuration (at least one).

The groups are:
* control-plane - which is either openstack-control or k8s-master
* data-plane - which is either openstack-compute or k8s-node

The role list which will be added for a given node depends
on to which of the two groups is the node assigned and on
the cloud_orchestrator value.

Change-Id: I58e98816bdbea6a38f7f33ffff3b4cff42c7f9aa