Make it possible to upgrade tox
As we moved pip/virtualenv/tox installation from nodepool
to ansible we are currenlty installing an older version
of tox. This makes running UT's on centos fail. We don't
want to upgrade it globally for now since it might break
other parts of the system. This makes it possible to decide
whether to upgrade tox when including the role for all
systems.

Change-Id: I7fffc2883f407854fceaff5d570eeff3365fb92f