Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> why these packages was added to environment instead of adding them to list of dependencies of package that requires them?

It was added part of the environment since all modules are going through py3 conversion need this. It’s better to part of the base environment. Anyway, this is a temporary change and can be purged
once we are done with coversion process.

Patch-set: 1
Reviewer: Gerrit User 526 <526@c72cd27d-1743-4a55-84a4-cd13b8400c52>