Update patch set 3
Patch Set 3:

This change depends on a change to a config project.

The syntax of the configuration in this change has been verified to be
correct once the config project change upon which it depends is
merged, but it can not be used until that occurs.

For more information about the OpenContrail CI system, see:
https://github.com/Juniper/contrail-project-config/wiki/OpenContrail-CI
Job live status page: http://zuulv3.opencontrail.org/

Patch-set: 3
Reviewer: Gerrit User 483 <483@c72cd27d-1743-4a55-84a4-cd13b8400c52>